google-translate-backtranslation-da/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
et-xmlfile==1.0.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-translate==1.7.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
jdcal==1.4.1
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
xlrd==1.2.0
XlsxWriter==1.2.8
